+++ src/mainui/sidebar/publicationwidget.cpp
@@ -473,9 +473,9 @@ void PublicationWidget::saveAnnotationContent()
     Nepomuk2::setProperty(resUri, NIE::htmlContent(), value);
 }
 
-void PublicationWidget::changeRating(int newRating)
+void PublicationWidget::changeRating(uint newRating)
 {
-    if(newRating == (int)m_publication.rating() ) {
+    if(newRating == m_publication.rating() ) {
         return;
     }
 
@@ -583,7 +583,7 @@ void PublicationWidget::setupWidget()
     ui->editPubMed->setPropertyUrl( NBIB::pubMed() );
     ui->editDOI->setPropertyUrl( NBIB::doi() );
 
-    connect(ui->editRating, SIGNAL(ratingChanged(uint)), this, SLOT(changeRating(int)));
+    connect(ui->editRating, SIGNAL(ratingChanged(uint)), this, SLOT(changeRating(uint)));
     connect(ui->editAnnot, SIGNAL(selectedAnnotation(Nepomuk2::Resource&)), this, SLOT(newAnnotationSelected(Nepomuk2::Resource&)));
 }
